Move deferred controls to interface
